Surface the git graph context menu in the history tab
